backend/requirements.txt
azure-core
azure-cosmos
azure-functions
azure-storage-blob
certifi
cffi
charset-normalizer
cryptography
docx2txt
idna
isodate
MarkupSafe
pycparser
requests
six
typing_extensions
urllib3
Werkzeug
azure-servicebus
azure-identity
azure-keyvault-secrets
beautifulsoup4
openai
pypdf
python-docx
scikit-learn
numpy
Jinja2
aiohttp
python-dotenv